Exchanged order of definitions of frame-manager and frame classes to avoid compile-time message.
Removed unused definition of calculate-standard-panes.

Changed order between adopt-child and compose-space.  Now, the mirrors of
the sheets in the pane hierarchy are realized first.  Their sizes are
arbitrarily assigned, since the space allocation protocol has not been
invoked yet.  Then compose-space and allocate-space are run to adjust all
the sizes.  This order allows us to determine space requirements of certain
panes from their contents, which requires the pane to be grafted so that
font sizes are known.