Redefined text-style-{ascent,descent,height,width} to conform to specification and changed all invocations.  Removed local variable `port' in :around functions for stream-write-xxx in recording.lisp as it is no longer necessary with the new definition of these functions.
There were two methods port-character-width on a clx-port used by
stream-character-width and character-width on a generic port (a
fall-through method) which was never used.  I renamed character-width
to port-character-width and made its signature identical to that of
port-character-width.

To be consistent with port-character-width above, I renamed the
generic function line-height, which is now called port-line-height.
There is currently only one method on it, the fall-through method on a
generic port, which is OK I guess since it is currently never called.