Replaced direct references to the slots text-style and line-style of medium (using slot-value or with-slot) by calls to the corresponding accessor.  Also added comments on these slots to make sure nobody uses them directly.
I am preparing for putting :after method on the setf methods on a
clx-medium.   These :after methods will update the graphics context.
That way, we will not have to update the graphics context every time
we call medium-gcontext the way we do now.

Currently, a large part of the time in the address-book demo is spent
updating the graphics context, and I want to speed it up.