Added :before methods to (setf medium-text-style) and (setf medium-line-style) that forward modifications to the graphics context when it exists.  The idea is to avoid modifying the graphics context as much as possible, since that turns out to be quite costly.  A similar thing should be done for other medium components such as the ink, clipping region, etc.
Modified function medium-gcontext to set the fields in the graphics
context having to do with line-style and text-style only when a new
graphics context is created.  Also fixed Julien's comment by adding a
semicolon so that it would indent properly.

The address-book example now displays faster, but not fast enough.
The problem is that replay-output-record always creates a new medium.
Doing so has two major consequences.  First, it slows down the
medium-gcontext function because there is never a graphics context
allocated so it has to be allocated and filled in for each call.
Second (worse) it does not deallocate the graphics contexts.  If we
are lucky, CLX will do that for us.  If not, we will eventually run
out of X resources.

The main time consumer in the address book example is now
port-lookup-mirror (around 10%).  The current method (a hash table in
the port) should be replaced by one in which the mirror is a field in
the mirrored sheet.

Fixed a spelling error in recent comments in the medium class and
signed the comments.

Started signing and dating my comments (as recommended by Norvig),
since I discovered I needed to know who had written a comment that was
not signed.  I used ISO date format to avoid international
misunderstandings.