Added method for sheet-native-transformation on top-level-sheet-pane which returns +identity-tranformation+.
The top leve sheet is now used as a reference for the native
transformation rather than the graft.  This way, we avoid strange
behavior when the window manager reparents the top level sheet.

Later, we may want the frame manager to decide whether the top level
sheet should have the identity tranformation or some other
transformation (presumably a scaling).