I removed the quoted expression in (setf gadget-current-color) so that colored gadgets change immediately when their color is altered.  I still don't understand why this expression was quoted.
I added standard-space-requirement-options-mixin to the list of
superclasses of menu-button-pane, and added a method for compose-space
on menu-button-pane.  Now, menu buttons have roughly the right size
again but the don't yet change colors according to whether they are
armed or not.  I suspect it is a question of adding
enter/exit-arms/disarms-mixin to the list of superclasses, but I
haven't had time to do that yet.

Now, there is considerable code duplication between the compose-space
methods of push-button and menu-button.  I am not sure how to factor
this yet.