Minor rewriting of the cursor attribute functions to conform more closely to the spec.
CURSOR-ACTIVE and CURSOR-STATE were not previously implemented.
CURSOR-VISIBILITY did not properly distinguish between the :on and :off
keywords.

The "visibility" slot of cursor-mixin has been eliminated, and replaced
with two slots, cursor-active and cursor-state. (setf cursor-visibility) has been
rewritten to modify these as appropriate, and CURSOR-VISIBILITY looks at
these to decide its return value.

The handling of cursor-active is likely not quite right, as I mostly ignore it.
The spec is not clear on what the "activeness" of a cursor means, but I assume
that deactivating a cursor should cause it to be erased from the display, which
I'm not doing presently. Presumably, this should not affect the cursor-state.
Presently no one uses CURSOR-ACTIVE.