In input-editor-format and %replace-input test for rescanning with stream-rescanning-p, not a bogus test of the scan pointer.
Comment out accept-using-read in Listener/hotfixes.lisp. I realize
we're losing some of the error-handling behavior there, but a bad side
effect of calling read-preserving-whitespace directly on the input
editing stream is that r-p-w calls (accept 'form), so all of sorts of
presentation translators become applicable unexpectedly. With the
patched accept-using-read commented out the graph nodes in the
listener are now acceptable as input to the "Show" commands.