Implement the :INITIALLY-SELECT-QUERY-IDENTIFIER option of ACCEPTING-VALUES. Update the text entry field in ACCEPTING-VALUES when the :DEFAULT argument in the corresponding call to ACCEPT changes. Return the correct third value from ACCEPT when the user enters new input.
Add some documentation about INCREMENTAL-DISPLAY and the difference
set.

Added FIND-PRESENTATION-TYPE-CLASS and CLASS-PRESENTATION-TYPE-NAME
which are trivial but which were not implemented.

Examples/accepting-values.lisp is a couple of examples from the Franz
user guide.

Tools/unimplemented.lisp prints out all the symbols in the CLIM
package which have no definition.