The old version was probably not portable, it's now using the "portablized" version LIST-DIRECTORY that is defined in util.lisp. More important a wild pathname will be created first.
(This actually has bitten me: I've created a core image that includes
 my File Selector that uses the functions of the Clim Listener to
 display directories. This image was going to be run on other
 computers that have no mcclim sources installed (not to speak about
 installed in equivalent directories). As PRECACHE-ICONS did not work
 and the icons for which there are no separate DEFICON constructs have
 not been cached, it tried to load the icon "up-folder.xpm" in
 COM-SHOW-DIRECTORY. . .)


By the way, what do you think about getting rid of the pathname
functions in util.lisp and depend on CL-FAD instead (the, so to speak,
maintained version of Peter Seibel's :com.gigamonkeys.pathnames
package)?  Why shouldn't we use a nice library if it already exists?
(In addition to this, the File Selector uses it as well --
particularly cl-fad:pathname-as-directory --, although it is not yet
commited to the CVS.)