Enable :pointer-motion in default event mask for realize-mirror-aux.  This unbreaks tracking-pointer's pointer-motion for all regular sheets. Previously, it only worked for clim-stream-pane, because clim-stream-pane overrode the defaults.
This will cause additional traffic between the X11
server and CLIM, however the CLIM spec does not provide a way to
selectively enable or disable events (only total muting is
possible). Therefore, we must shove all events to a sheet as we have
no way to find out if an event is handled or not.