Add highly experimental Null backend.
The idea is that the null backend implements all the mcclim machinery
for a backend, but doesn't side-effect the rest of the world; this
should make it possible to write test cases for mcclim-internal
invariants, and potentially also mcclim applications, by running them
under this backend.  This utopia is quite a way off, however; what
actually works at present is not much more than:
  (setf clim:*default-server-path* :null)
  (let ((stream (clim:open-window-stream)
    (clim:draw-rectangle* stream 10 10 100 200)
    (clim:stream-output-history stream))
but it's a start.

(Additionally, the Null backend could be used as a starting point for
implementing other backends.)