Fixed the `compose-space' method for Drei-based text-editor-panes to suck less and be more correct. The :nlines value will still not work properly due to McCLIM layouting brokenness.
Also changed the `text-editor' gadget to use a completely standard
Drei gadget-pane, without minibuffer, scrollbars and anything
else. This is necessary because we need to return the actual pane
object from the function, so we can't wrap it in nice things like
borders, scrollbars or a minibuffer. Use the :drei gadget for the nice
version.