Added in-line completion using input-editor typeout instead of calling menu-choose.
Does not work in Goatee.

May fail under some circumstances that input-editor typeout doesn't handle well yet.

May behave illogically because the list of completions is kept alive
for fairly long, yet the presentations on it become untouchable almost
immediately. This is because they are of a specially created
completion presentation type, and not the more general presentation
type of the object they represent. This knowledge is not accessible to
the input-editing machinery (also, it seems presentation type options
are compared for equality using EQ/EQL, so two content-wise identical
possibility-lists can have different completion presentation types).